Support ActiveModel v3.2.12
ActiveModel v3.2.12 has a bug on `Object#dup`.
See https://github.com/rails/rails/pull/8405 for more details.

I decided to use `Object#clone` instead of `Object#dup`,
while this is a dirty hack, but anything works fine with this.

In addition, I replaced `ActiveModel#validate` with `#valid?`
because it does not exist on v3.